refactor(backend): split message repository responsibilities
